Change scope selection in the modify permissions UI (#165)
* Change how we show which scopes have been selected.
* Disabled checkboxes on scopes that have already gotten consent, track whether we have any admin scopes selected.
* Added loc string

Before, if a user had provided consent for a scope, the scope was selected in the checkbox in the UI. This caused issues since the same checkbox mechanism was used to request scopes. This led to some confusing UI behavior. We now show consented scopes with the word 'Consented' next to the scope name when a user has consented to that scope. The check boxes are now only to show that we want to select a scope. We also disable scopes that have been previously consented since there is no reason to select them again.

This also gets rid of the issue where if a user had previously received admin consent, every time that they wanted to consent to an additional user scope, they'd have to deselect the admin scopes in order to not get stuck in the admin consent flow.